MVP FINAL COMMIT Everything working meets all requirements
