--- CALORIE-TRACKER/CALORIE-TRACKER/CalorieIntake.xcdatamodeld/CalorieIntake.xcdatamodel/contents
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<model type="com.apple.IDECoreDataModeler.DataModel" documentVersion="1.0" lastSavedToolsVersion="16119" systemVersion="19E287" minimumToolsVersion="Automatic" sourceLanguage="Swift" userDefinedModelVersionIdentifier="">
-    <entity name="CalorieIntake" representedClassName="CalorieIntake" syncable="YES" codeGenerationType="class">
-        <attribute name="identifier" attributeType="String"/>
-        <attribute name="numOfCalories" attributeType="String"/>
-        <attribute name="timeStamp" attributeType="Date" usesScalarValueType="NO"/>
-    </entity>
-    <elements>
-        <element name="CalorieIntake" positionX="-63" positionY="-18" width="128" height="88"/>
-    </elements>
-</model>
