--- CALORIE-TRACKER/CALORIE-TRACKER/CalorieIntakeTableViewCell.swift
@@ -1,36 +1,0 @@
-//
-//  CalorieIntakeTableViewCell.swift
-//  CALORIE-TRACKER
-//
-//  Created by Kelson Hartle on 6/19/20.
-//  Copyright © 2020 Kelson Hartle. All rights reserved.
-//
-
-import UIKit
-
-class CalorieIntakeTableViewCell: UITableViewCell {
-    
-    @IBOutlet weak var calorieIntakeLabel: UILabel!
-    @IBOutlet weak var timeStampLabel: UILabel!
-    
-
-    
-    var calorieIntake: CalorieIntake? {
-        didSet {
-            updateViews()
-        }
-    }
-    
-    private func updateViews() {
-        guard let calorieIntake = calorieIntake else { return }
-        
-        let dateFormatter = DateFormatter()
-        
-        dateFormatter.dateFormat = "MM/dd/YY, h:mm a"
-        let now = dateFormatter.string(from: calorieIntake.timeStamp!)
-        
-        calorieIntakeLabel.text = calorieIntake.numOfCalories
-        timeStampLabel.text = now
-    }
-
-}
