+++ CALORIE-TRACKER/CALORIE-TRACKER/Notifications/Notifications.swift
@@ -1,0 +1,14 @@
+//
+//  Notifications.swift
+//  CALORIE-TRACKER
+//
+//  Created by Kelson Hartle on 6/19/20.
+//  Copyright © 2020 Kelson Hartle. All rights reserved.
+//
+
+import Foundation
+
+extension NSNotification.Name {
+    static let caloriesHaveBeenAdded = NSNotification.Name("caloriesHaveBeenAdded")
+
+}
